requirements.txt
# Vineyard Group Fellowship Django Backend - Production Requirements
# This file contains all dependencies needed for production deployment
# Install with: pip install -r requirements.txt

# === CORE DJANGO FRAMEWORK ===
Django==5.2.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1

# === DATABASE & CACHING ===
psycopg==3.2.10
psycopg-binary==3.2.10
redis==5.0.1
dj-database-url==2.1.0

# === AUTHENTICATION & SECURITY ===
django-cors-headers==4.3.1
django-otp==1.3.0
django-fernet-fields==0.6
PyJWT==2.10.1
cryptography==46.0.2
argon2-cffi==23.1.0

# === CONTENT SECURITY & RATE LIMITING ===
django-csp==3.8
django-ratelimit==4.1.0

# === API DOCUMENTATION ===
drf-spectacular==0.27.2
drf-nested-routers==0.93.5

# === EMAIL SERVICES ===
django-anymail==10.3
sendgrid==6.11.0

# === IMAGE PROCESSING ===
pillow==10.4.0

# === UTILITIES ===
python-decouple==3.8
python-dateutil==2.9.0.post0
user-agents==2.2.0
ua-parser==1.0.1

# === MONITORING & LOGGING ===
sentry-sdk[django]==1.45.0
structlog==23.2.0

# === WEBSOCKETS (for future real-time features) ===
channels==4.0.0
channels-redis==4.2.0

# === PRODUCTION WSGI SERVER ===
gunicorn==21.2.0

# === STATIC FILES SERVING ===
whitenoise==6.6.0

# === PRODUCTION SECURITY ===
bandit==1.7.5
pip-audit==2.6.2